requirements.txt
alembic==1.18.4
APScheduler==3.11.2
bcrypt==4.3.0
fastapi==0.135.0
httpx==0.28.1
passlib==1.7.4
psycopg2-binary==2.9.11
pydantic==2.12.5
pydantic-settings==2.13.1
PyJWT==2.11.0
python-dotenv==1.2.1
python-multipart==0.0.22
SQLAlchemy==2.0.47
uvicorn==0.41.0
pytest==9.0.2